+++ itsys_real_estate/models/ownership_contract.py
@@ -67,7 +67,7 @@ class ownership_contract(models.Model):
     city= fields.Many2one('cities','City')
     user_id= fields.Many2one('res.users','Responsible', default=lambda self: self.env.user,)
     partner_id= fields.Many2one('res.partner','Partner',required=True)
-    building_area= fields.Integer ('Building Unit Area m²',)
+    building_area = fields.Float ('Building Unit Area m²',)
     loan_line= fields.One2many('loan.line.rs.own', 'loan_id')
     region= fields.Many2one('regions','Region')
     country= fields.Many2one('countries','Country')
